Refactor: Move Icons into Utils folder
